Allow user to specify table.
There are use cases where the table is not located at
`$scope.$parent.ngTableParams`.

One example is when using Angular's controllerAs syntax. In this case,
things are not usually stored on `$scope`.

This commit will not break existing uses of `ngTableExport` because the
code defaults to `$scope.$parent.table` when no `table` arugment is
provided.